Add a vagrant setup for testing
I primarily use a Macbook Air. It's hard to test scripts which only run on
Linux on this platform. Vagrant is a nice way of supporting running an
arbitrary VM for this script and I've added one that is slightly configurable
depending on whether you prefer RedHat (actually CentOS) or Debian (actually
Ubuntu)