Make almost everything optional
One of the goals Alex had, which I also wanted to do back in the day, is make the
checks much more optional, so I have added a very basic option handling setup to
the script.

By default, all options are enabled (to maintain expected usage), but you can
turn off everything by the OS reporting with `-n` and then turn on the individual
reports with their respective flags (c=CPU, r=RAM, d=Disk, p=PHP, m=MySQL).

This has the added advantage of being able to run *some* of this script on servers
without sysstat installed or in environments where you should not install packages
by only checking for the sysstats package if CPU or RAM stats are going to be
displayed.

This then opens up the ability to have some more functions in the script without
changing the default output, as they could default to being turned off unless
enabled with a flag.