Rejig how Vagrant works
This removes the need for environment variables. By default
vagrant will start one VM (`centos`) and that will be the machine
used for all vagrant commands. In addition you can bring up `ubuntu` with
`vagrant up ubuntu`, but you will need to supply `ubuntu` as an argument
to all vagrant commands. The `centos` VM can continue to be used without
an argument as if there was only one machine.